Addiflare/requirements.txt
aniso8601==7.0.0
asgiref==1.1.2
attrs==19.3.0
autobahn==19.10.1
Automat==0.7.0
cffi==1.13.0
channels==1.1.8
channels-api==0.4.1
constantly==15.1.0
cryptography==2.8
daphne==1.4.2
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
django-graphql-jwt==0.2.2
djangorestframework==3.10.3
graphene==2.1.8
graphene-django==2.5.0
graphene-django-extras==0.4.7
graphene-django-subscriptions==0.0.6
graphql-core==2.2.1
graphql-relay==2.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
iso8601==0.1.12
promise==2.2.1
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
zope.interface==4.6.0
